Engine graphics library wrapper

+++ libs/Packet/PacketManager.cpp
@@ -25,45 +25,91 @@ PacketManager::PacketManager(uint32_t seqNumber) {
     this->_payload = std::vector<uint64_t>();
     this->_serializer = std::make_shared<BigEndianSerialization>();
     if (!this->_serializer) {
-        throw err::PacketError("[Server] Failed to load serializer",
-            err::PacketError::SERIALIZER_ATTRIBUTION_FAILED);
+        throw err::PacketError(
+            "[Server] Failed to load serializer",
+            err::PacketError::SERIALIZER_ATTRIBUTION_FAILED
+        );
     }
 
     this->_packetHandlers = {
-        {CONNECTION_CLIENT_PACKET, std::bind(
-            &PacketManager::buildConnectionPacket,
-            this, std::placeholders::_1)},
-        {ACCEPTATION_PACKET, std::bind(
-            &PacketManager::buildAcceptationPacket,
-            this, std::placeholders::_1)},
-        {DISCONNECTION_PACKET, std::bind(
-            &PacketManager::buildDisconnectionPacket,
-            this, std::placeholders::_1)},
-        {EVENT_PACKET, std::bind(
-            &PacketManager::buildEventPacket,
-            this, std::placeholders::_1)}
+        {
+            static_cast<uint8_t>(CONNECTION_CLIENT_PACKET),
+            std::bind(
+                &PacketManager::buildConnectionPacket,
+                this, std::placeholders::_1
+            )
+        },
+        {
+            static_cast<uint8_t>(ACCEPTATION_PACKET),
+            std::bind(
+                &PacketManager::buildAcceptationPacket,
+                this, std::placeholders::_1
+            )
+        },
+        {
+            static_cast<uint8_t>(DISCONNECTION_PACKET),
+            std::bind(
+                &PacketManager::buildDisconnectionPacket,
+                this, std::placeholders::_1
+            )
+        },
+        {
+            static_cast<uint8_t>(EVENT_PACKET),
+            std::bind(
+                &PacketManager::buildEventPacket,
+                this, std::placeholders::_1
+            )
+        }
     };
 
     this->_packetReceived = {
-        {CONNECTION_CLIENT_PACKET, std::bind(
-            &PacketManager::parseConnectionPacket,
-            this, std::placeholders::_1)},
-        {ACCEPTATION_PACKET, std::bind(
-            &PacketManager::parseAcceptationPacket,
-            this, std::placeholders::_1)},
-        {DISCONNECTION_PACKET, std::bind(
-            &PacketManager::parseDisconnectionPacket,
-            this, std::placeholders::_1)},
-        {EVENT_PACKET, std::bind(
-            &PacketManager::parseEventPacket,
-            this, std::placeholders::_1)}
+        {
+            static_cast<uint8_t>(CONNECTION_CLIENT_PACKET),
+            std::bind(
+                &PacketManager::parseConnectionPacket,
+                this, std::placeholders::_1
+            )
+        },
+        {
+            static_cast<uint8_t>(ACCEPTATION_PACKET),
+            std::bind(
+                &PacketManager::parseAcceptationPacket,
+                this, std::placeholders::_1
+            )
+        },
+        {
+            static_cast<uint8_t>(DISCONNECTION_PACKET),
+            std::bind(
+                &PacketManager::parseDisconnectionPacket,
+                this, std::placeholders::_1
+            )
+        },
+        {
+            static_cast<uint8_t>(EVENT_PACKET),
+            std::bind(
+                &PacketManager::parseEventPacket,
+                this, std::placeholders::_1
+            )
+        }
     };
 
     this->_packetLengths = {
-        {CONNECTION_CLIENT_PACKET, LENGTH_CONNECTION_PACKET},
-        {ACCEPTATION_PACKET, LENGTH_ACCEPTATION_PACKET},
-        {DISCONNECTION_PACKET, LENGTH_DISCONNECTION_PACKET},
-        {EVENT_PACKET, LENGTH_EVENT_PACKET}
+        {
+            static_cast<uint8_t>(CONNECTION_CLIENT_PACKET),
+            LENGTH_CONNECTION_PACKET
+        },
+        {
+            static_cast<uint8_t>(ACCEPTATION_PACKET),
+            LENGTH_ACCEPTATION_PACKET
+        },
+        {
+            static_cast<uint8_t>(DISCONNECTION_PACKET),
+            LENGTH_DISCONNECTION_PACKET
+        },
+        {
+            static_cast<uint8_t>(EVENT_PACKET),
+            LENGTH_EVENT_PACKET
+        }
     };
 }
